Provide indication of time travel form validation error
