+++ src/components/editor/Bindings/AdvancedOptions/types.tsx
@@ -1,0 +1,9 @@
+export interface AdvancedOptionsProps {
+    bindingIndex: number;
+    bindingUUID: string;
+    collectionName: string;
+}
+
+export interface HeaderProps {
+    errorsExist: boolean;
+}
