+++ src/components/editor/Bindings/types.ts
@@ -1,11 +1,5 @@
 export type TabOptions = 'config' | 'schema' | 'field_selection';
 
-export interface AdvancedOptionsProps {
-    bindingIndex: number;
-    bindingUUID: string;
-    collectionName: string;
-}
-
 export interface CollectionData {
     spec: any;
     belongsToDraft: boolean;
